Prevent product quantity reset when leaving Shop
To do this, the allProductsData state needed to be set in RouteSwitch so
the data were not update each time the Shop was navigated to. When it
was in the Shop, the product ids were updating each time the Shop was
opened so the logic to check if the shoppingCart contained the products
by id was not finding the newly generate ids. Therefore, the quantities
would be set to 0 again.